Quinny/requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.3
bootstrap4==0.1.0
certifi==2020.6.20
chardet==3.0.4
dj-database-url==0.5.0
dj-email-url==1.0.1
Django==3.1.2
django-bootstrap4==2.3.1
django-cache-url==3.1.2
django-cors-headers==3.5.0
django-crispy-forms==1.9.2
django-filter==2.4.0
django-tempus-dominus==5.1.2.13
djangorestframework==3.12.1
environs==8.0.0
gunicorn==20.0.4
idna==2.10
importlib-metadata==2.0.0
lxml==4.6.1
marshmallow==3.9.0
numpy==1.19.3
pandas==1.1.4
psycopg2-binary==2.8.6
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
urllib3==1.25.11
whitenoise==5.2.0
zipp==3.4.0
